+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Next
+- Add the ability to smooth raw signals using convolution with a Gaussian or flat kernel. 
+- Add the ability to smooth raw signals using a Savitzky-Golay filter
+
 ## 0.2.4 - 2021-10-13
 
 ### Added